Tweaks to the textBox
Fixed bugs and polished the PDF save functionality.

Created a new module for the save functions.

Organized the textBox title to be un-editable and has its own styling.

Fixed the focus() when textBox is created so the caret always goes to the start of the new line.

All new text input in textBox is <p> not <div> anymore in Chrome.
