+++ index.html
@@ -136,7 +136,8 @@
                 <div class="text-area">
                         <div data-txt-area="text-area" 
                         class="textbox" 
-                        contenteditable="true"><p data-txt-name="name-textbox"></p><p>//</p></div>
+                        contenteditable="true"><p data-txt-name="name-textbox" 
+                        contenteditable="false" class="text-area-title font-size-sm"></p><p>&nbsp</p></div>
                 </div>
         </div>
 </div>
