+++ scripts/jsconfig.json
@@ -10,5 +10,5 @@
     "dictionary/*.js",
     "color/*.js",
     "c:\\Users\\USUARIO\\.vscode\\extensions\\samplavigne.p5-vscode-1.2.8\\p5types\\global.d.ts", 
-    "modal/*.js", "textBox/*.js", "textBox/wordCount/dictionary/dictionary.js", "textBox/wordCount/graphics/sketch.js", "textBox/wordCount/graphics/colorChoose.js"]
+    "modal/*.js", "textBox/*.js", "textBox/wordCount/dictionary/*.js", "textBox/wordCount/graphics/*.js"]
 }
