+++ scripts/main.js
@@ -19,5 +19,3 @@ dragModal();
 modal.createModalBtns();
 textBox.createTxtBtns();
 textBox.updateNameOfText();
-
-
