+++ scripts/textBox/textBox.js
@@ -34,6 +34,4 @@ export default class TextBox{
         });
              
     };
-
-
 };
