+++ scripts/textBox/textOptions.js
@@ -1,3 +1,6 @@
+import { textArea } from "../elements/variablesDOM.js";
+import { saveTxt, export2PDF, export2Word } from "./saveText.js";
+
 export function colorSelect(btn, textArea){
     textArea.focus();
     let sel = window.getSelection().getRangeAt(0).commonAncestorContainer.parentElement;
@@ -37,20 +40,17 @@ function formatBtns(btn, textArea){
     if(!checkSelection()) return;
 
     let btnFormat = btn.getAttribute('data-format');
-    let selection = window.getSelection().getRangeAt(0);
-    console.log("Selection: " + selection + ' ' + selection.startOffset + ' ' + selection.endOffset);
-    console.log(selection);
 
     textArea.focus();
     switch (btnFormat){
         case 'bold':
-
+            document.execCommand('bold', false);
             break;
         case 'italic':
-
+            document.execCommand('italic', false);
             break;
         case 'underline':
-
+            document.execCommand('underline', false);
             break;
     }
 
@@ -59,14 +59,7 @@ function formatBtns(btn, textArea){
 //Function to align-set size of btns
 function dropBtns(btn, type){
 
-    let selection;
-
-    console.log(checkSelection());
-
     if(!checkSelection()) return;
-
-    selection = window.getSelection().getRangeAt(0).commonAncestorContainer.parentElement;
-    
 
     switch(type){
         case 'align': {
@@ -81,104 +74,16 @@ function dropBtns(btn, type){
     };
 };
 
-//Function to save the text TXT
-function saveTxt(textArea, fileName){
-
-    const a = document.createElement('a');
-    const blob = new Blob([textArea.innerText]);
-    const dataUrl = URL.createObjectURL(blob);
-    a.href = dataUrl;
-    a.download = fileName + '.txt';
-    a.click();
-
-};
-
-function export2Word(textArea, filename){
-    const preHtml = "<html xmlns:o='urn:schemas-microsoft-com:office:office' xmlns:w='urn:schemas-microsoft-com:office:word' xmlns='http://www.w3.org/TR/REC-html40'><head><meta charset='utf-8'><title>Export HTML To Doc</title></head><body>";
-    const postHtml = "</body></html>";
-    const html = preHtml+textArea.innerHTML+postHtml;
-
-    var blob = new Blob(['\ufeff', html], {
-        type: 'application/msword'
-    });
-    
-    // Specify link url
-    var url = 'data:application/vnd.ms-word;charset=utf-8,' + encodeURIComponent(html);
-    
-    // Specify file name
-    filename = filename?filename+'.doc':'document.doc';
-    
-    // Create download link element
-    var downloadLink = document.createElement("a");
-
-    document.body.appendChild(downloadLink);
-    
-    if(navigator.msSaveOrOpenBlob ){
-        navigator.msSaveOrOpenBlob(blob, filename);
-    }else{
-        // Create a link to the file
-        downloadLink.href = url;
-        
-        // Setting the file name
-        downloadLink.download = filename;
-        
-        //triggering the function
-        downloadLink.click();
-    }
-    
-    document.body.removeChild(downloadLink);
-};
-
-function export2PDF(textArea, fileName){
-    //Little tweak for the JSPDF to work:
-    window.jsPDF = window.jspdf.jsPDF;
-
-    //Creating the jsPDF object with some properties:
-    const doc = new jsPDF({
-        orientation: 'p',
-        unit: 'mm',
-        format: 'letter'
-    });
-
-    //Created a splitTextToSize array, it takes the whole textarea.inner text string and divides it
-    //into an array of strings according to the size put in the second arg.
-    const splitTxt = doc.splitTextToSize(textArea.innerText, 250);
-
-    const pageRender = (i) =>{
-        doc.setFontSize(8);
-        let page = "Page: " + (i+1);
-        doc.setPage(i+1);
-        doc.text(page, doc.getPageWidth() - 30, doc.getPageHeight() - 10);
-    };
-
-    let y = 20;
-    let p = 0;
-    doc.setFontSize(8);
-    doc.text("Created using Wordscope", 5, 5);
-    for(let i = 0; i < splitTxt.length; i++){
-        pageRender(i);
-        if(y > doc.getPageHeight() - 40){
-            y = 15;
-            doc.addPage();
-        }
-        doc.setFontSize(12);
-        doc.text(splitTxt[i], 10, y);
-        y += 5;
-    }
-
-    doc.save(fileName + ".pdf");
-};
-
 function checkSelection(){
 
     let result;
 
     if(window.getSelection().anchorNode === null) return false;
     
-    let isTextArea = window.getSelection().focusNode.parentElement.parentNode;
-    
-    if(isTextArea.getAttribute('class') === 'text-area' || isTextArea.getAttribute('class') === 'textbox') result = true;
-    else result = false;
+    //DOCUMENT_POSITION_CONTAINS and DOCUMENT_POSITION_PRECEDING value is 10 - meaning the element is on textBox
+    //Check the documentation on compareDocumentPosition()
+    let isTextArea = window.getSelection().focusNode.compareDocumentPosition(textArea);
+    isTextArea === 10 ? result = true : result = false;
 
     return result;
 };
